+++ src/zeal/listeners.py
@@ -1,4 +1,6 @@
+import inspect
 import logging
+import warnings
 from abc import ABC, abstractmethod
 from collections import defaultdict
 from contextlib import contextmanager
@@ -10,7 +12,7 @@ from typing import Optional, TypedDict
 from django.conf import settings
 from django.db import models
 
-from zeal.util import get_caller
+from zeal.util import get_caller, get_stack
 
 from .errors import NPlusOneError, ZealError
 
@@ -32,8 +34,9 @@ class AllowListEntry(TypedDict):
 
 @dataclass
 class NPlusOneContext:
-    counts: dict[CountsKey, int] = field(
-        default_factory=lambda: defaultdict(int)
+    enabled: bool = False
+    calls: dict[CountsKey, list[list[inspect.FrameInfo]]] = field(
+        default_factory=lambda: defaultdict(list)
     )
     ignored: set[str] = field(default_factory=set)
     allowlist: list[AllowListEntry] = field(default_factory=list)
@@ -64,9 +67,20 @@ class Listener(ABC):
 
         return [*settings_allowlist, *_nplusone_context.get().allowlist]
 
-    def _alert(self, model: type[models.Model], field: str, message: str):
+    def _alert(
+        self,
+        model: type[models.Model],
+        field: str,
+        message: str,
+        calls: list[list[inspect.FrameInfo]],
+    ):
         should_error = (
             settings.ZEAL_RAISE if hasattr(settings, "ZEAL_RAISE") else True
+        )
+        should_include_all_callers = (
+            settings.ZEAL_SHOW_ALL_CALLERS
+            if hasattr(settings, "ZEAL_SHOW_ALL_CALLERS")
+            else False
         )
         is_allowlisted = False
         for entry in self._allowlist:
@@ -81,12 +95,24 @@ class Listener(ABC):
         if is_allowlisted:
             return
 
-        caller = get_caller()
-        message = f"{message} at {caller.filename}:{caller.lineno} in {caller.function}"
+        final_caller = get_caller()
+        if should_include_all_callers:
+            message = f"{message} with calls:\n"
+            for i, caller in enumerate(calls):
+                message += f"CALL {i+1}:\n"
+                for frame in caller:
+                    message += f"  {frame.filename}:{frame.lineno} in {frame.function}\n"
+        else:
+            message = f"{message} at {final_caller.filename}:{final_caller.lineno} in {final_caller.function}"
         if should_error:
             raise self.error_class(message)
         else:
-            logger.warning(message)
+            warnings.warn_explicit(
+                message,
+                UserWarning,
+                filename=final_caller.filename,
+                lineno=final_caller.lineno,
+            )
 
 
 class NPlusOneListener(Listener):
@@ -101,13 +127,15 @@ class NPlusOneListener(Listener):
         instance_key: Optional[str],
     ):
         context = _nplusone_context.get()
+        if not context.enabled:
+            return
         caller = get_caller()
         key = (model, field, f"{caller.filename}:{caller.lineno}")
-        context.counts[key] += 1
-        count = context.counts[key]
+        context.calls[key].append(get_stack())
+        count = len(context.calls[key])
         if count >= self._threshold and instance_key not in context.ignored:
             message = f"N+1 detected on {model.__name__}.{field}"
-            self._alert(model, field, message)
+            self._alert(model, field, message, context.calls[key])
         _nplusone_context.set(context)
 
     def ignore(self, instance_key: Optional[str]):
@@ -138,7 +166,9 @@ def setup() -> Optional[Token]:
     # if we're already in an ignore-context, we don't want to override
     # it.
     context = _nplusone_context.get()
-    return _nplusone_context.set(NPlusOneContext(allowlist=context.allowlist))
+    return _nplusone_context.set(
+        NPlusOneContext(enabled=True, allowlist=context.allowlist)
+    )
 
 
 def teardown(token: Optional[Token] = None):
@@ -164,7 +194,8 @@ def zeal_ignore(allowlist: Optional[list[AllowListEntry]] = None):
 
     old_context = _nplusone_context.get()
     new_context = NPlusOneContext(
-        counts=old_context.counts.copy(),
+        enabled=old_context.enabled,
+        calls=old_context.calls.copy(),
         ignored=old_context.ignored.copy(),
         allowlist=[*old_context.allowlist, *allowlist],
     )